build(gulp): Switch to gulp for building target formats
The gulpfile builds ES, CJS & UMD (unminified & minified). Also added `src/index.js` which will serve as an aggregator file so that imports can happen straight from the package, if desired. As a result, there is a new `src/compose.js` file and the other files will also live directly in `src/` instead of in subfolders of `src/`. This is why the separate README files are now also in `src/`. They also reflect the two different approaches for importing.

Rollup was only creating bundles when really, we want to transpile each individual file so that they could potentially be used individually. The rollup bundle would be _ok_ if you had tree-shaking with ES modules,
but other than that you'd be getting a huge bundle and only using a few functions.
