+++ package.json
@@ -3,10 +3,10 @@
   "version": "0.0.1",
   "description": "A collection of higher-order components (HOCs) to easily create composite events in React components",
   "author": "Ben Ilegbodu <ben@benmvp.com>",
-  "main": "dist/react-composite-events.cjs.js",
-  "module": "dist/react-composite-events.es.js",
-  "jsnext:main": "dist/react-composite-events.es.js",
-  "browser": "dist/react-composite-events.umd.js",
+  "main": "dist/cjs/index.js",
+  "module": "dist/es/index.js",
+  "jsnext:main": "dist/es/index.js",
+  "browser": "dist/umd/index.js",
   "keywords": [
     "react",
     "react-native",
@@ -27,7 +27,7 @@
   "homepage": "https://github.com/benmvp/react-composite-events#readme",
   "license": "MIT",
   "scripts": {
-    "build": "rm -rf dist && rollup --config",
+    "build": "gulp",
     "format": "prettier-eslint --write 'src/**/*.js'",
     "lint": "eslint src",
     "precommit": "lint-staged && npm run validate",
@@ -77,15 +77,25 @@
     "babel-jest": "^20.0.3",
     "babel-plugin-external-helpers": "^6.22.0",
     "babel-plugin-transform-class-properties": "^6.24.1",
+    "babel-plugin-transform-es2015-modules-umd": "^6.24.1",
     "babel-preset-env": "^1.6.0",
     "babel-preset-react": "^6.24.1",
     "babel-preset-stage-3": "^6.24.1",
+    "del": "^3.0.0",
     "enzyme": "^2.9.1",
     "eslint": "^3.0.0",
     "eslint-config-eventbrite-react": "^5.0.0",
     "eslint-plugin-import": "^2.0.0",
     "eslint-plugin-jsx-a11y": "^6.0.0",
     "eslint-plugin-react": "^6.0.0",
+    "gulp": "^3.9.1",
+    "gulp-babel": "^6.1.2",
+    "gulp-debug": "^3.1.0",
+    "gulp-rename": "^1.2.2",
+    "gulp-replace": "^0.6.1",
+    "gulp-sourcemaps": "^2.6.0",
+    "gulp-uglify": "^3.0.0",
+    "gulp-util": "^3.0.8",
     "husky": "^0.14.3",
     "jest": "^20.0.4",
     "jest-enzyme": "^3.6.1",
@@ -93,11 +103,7 @@
     "prettier-eslint-cli": "^4.1.1",
     "react": "^15.6.1",
     "react-dom": "^15.6.1",
-    "react-test-renderer": "^15.6.1",
-    "rollup": "^0.45.2",
-    "rollup-plugin-babel": "^2.7.1",
-    "rollup-plugin-commonjs": "^8.0.2",
-    "rollup-plugin-node-resolve": "^3.0.0"
+    "react-test-renderer": "^15.6.1"
   },
   "dependencies": {}
 }
