+++ src/generic.md
@@ -28,6 +28,8 @@ withLongPress
 Creates an HOC for a composite event (named `onLongPress-*`) that is triggered when a user presses down on a component and doesn't release from that component or move out of it for a specified duration of time.
 
 ```js
+import {withLongPress} from 'react-composite-events'
+// or
 import {withLongPress} from 'react-composite-events/generic'
 
 const EnhancedTO = withLongPress(750)(TouchableOpacity)
@@ -51,6 +53,8 @@ export default MyComponent extends PureComponent {
 When no `duration` parameter is specified, then the value for this parameter is defaulted to `1250` (1.25 seconds). In this case the composite event prop is simply `onLongPress`.
 
 ```js
+import {withLongPress} from 'react-composite-events'
+// or
 import {withLongPress} from 'react-composite-events/generic'
 
 const EnhancedAnchor = withLongPress()('a')
@@ -91,6 +95,8 @@ withRemainReleased
 Creates an HOC for a composite event (named `onRemainReleased-*`) that is triggered when a user releases a component (after pressing it) and doesn't press that component again for a specified duration of time.
 
 ```js
+import {withRemainReleased} from 'react-composite-events'
+// or
 import {withRemainReleased} from 'react-composite-events/generic'
 
 const EnhancedTO = withRemainReleased(250)(TouchableOpacity)
@@ -113,6 +119,8 @@ export default MyComponent extends PureComponent {
 When no `duration` parameter is specified, then the value for this parameter is defaulted to `500` (500 milliseconds). In this case the composite event prop is simply `onRemainReleased`.
 
 ```js
+import {withRemainReleased} from 'react-composite-events'
+// or
 import {withRemainReleased} from 'react-composite-events/generic'
 
 const EnhancedTO = withRemainReleased()(TouchableOpacity)
@@ -149,6 +157,8 @@ withRemainFocused
 Creates an HOC for a composite event (named `onRemainFocused-*`) that is triggered when a component stays focused for a specified duration of time.
 
 ```js
+import {withRemainFocused} from 'react-composite-events'
+// or
 import {withRemainFocused} from 'react-composite-events/generic'
 
 const EnhancedTI = withRemainFocused(1000)(TextInput)
@@ -169,6 +179,8 @@ export default MyComponent extends PureComponent {
 When no `duration` parameter is specified, then the value for this parameter is defaulted to `500` (500 milliseconds). In this case the composite event prop is simply `onRemainFocused`.
 
 ```js
+import {withRemainFocused} from 'react-composite-events'
+// or
 import {withRemainFocused} from 'react-composite-events/generic'
 
 const EnhancedTI = withRemainFocused()(TouchableOpacity)
@@ -203,6 +215,8 @@ withRemainBlurred
 Creates an HOC for a composite event (named `onRemainBlurred-*`) that is triggered when a component stays blurred / unfocused for a specified duration of time.
 
 ```js
+import {withRemainBlurred} from 'react-composite-events'
+// or
 import {withRemainBlurred} from 'react-composite-events/generic'
 
 const EnhancedTI = withRemainBlurred(350)(TextInput)
@@ -223,6 +237,8 @@ export default MyComponent extends PureComponent {
 When no `duration` parameter is specified, then the value for this parameter is defaulted to `500` (500 milliseconds). In this case the composite event prop is simply `onRemainBlurred`.
 
 ```js
+import {withRemainBlurred} from 'react-composite-events'
+// or
 import {withRemainBlurred} from 'react-composite-events/generic'
 
 const EnhancedTI = withRemainBlurred()(TouchableOpacity)
